# xiaoyi/plugin_openapi_v6.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "优活 Agent API",
    "description": "面向独居老人的家庭协同式可信生活智能体。语言层只负责理解；办事、循环事务、情绪陪伴、记忆、健康、用药、位置安全与家庭协同由可审计代码控制；v5新增语音共识、目的绑定策略、可恢复Saga、离线同步、破窗访问与可验证证明；v6新增认知负荷治理、玻璃盒依赖校准、安全预演、受约束语义网关与真实用户实验工具链。",
    "version": "6.0.0"
  },
  "paths": {
...
    "/v6/profiles/{elder_id}": {
      "get": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_v6_profiles__elder_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Elder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_v6_profiles__elder_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Elder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v6/interaction/plan": {
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Interaction Plan",
        "operationId": "interaction_plan_v6_interaction_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/reliance/card": {
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Reliance Card",
        "operationId": "reliance_card_v6_reliance_card_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelianceCardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelianceCard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/actions/preview": {
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Safe Preview",
        "operationId": "safe_preview_v6_actions_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/tasks/{task_id}/glass-box": {
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Task Glass Box",
        "description": "Glass-box card and safe preview for one real task (design §4.3).",
        "operationId": "task_glass_box_v6_tasks__task_id__glass_box_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskGlassBoxRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskGlassBox"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v6/semantic/parse": {
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Semantic Parse",
        "operationId": "semantic_parse_v6_semantic_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticFrame"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/studies/sessions": {
      "get": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "List Study Sessions",
        "operationId": "list_study_sessions_v6_studies_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudySession"
                  },
                  "type": "array",
                  "title": "Response List Study Sessions V6 Studies Sessions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Create Study Session",
        "operationId": "create_study_session_v6_studies_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/studies/observations": {
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Add Study Observation",
        "operationId": "add_study_observation_v6_studies_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyObservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyObservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/studies/summary": {
      "get": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Study Summary",
        "operationId": "study_summary_v6_studies_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/comprehension/{elder_id}": {
      "get": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Comprehension",
        "description": "Observed teach-back outcomes driving the interaction governor.\n\nOutcome labels only - never what the elder said - so the family can see\nthat support is adapting without reading the conversation.",
        "operationId": "comprehension_v6_comprehension__elder_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Elder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Comprehension V6 Comprehension  Elder Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v6/speech/voice": {
      "get": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Speech Voice",
        "description": "Whether an offline neural voice is available; the client degrades if not.",
        "operationId": "speech_voice_v6_speech_voice_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Speech Voice V6 Speech Voice Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/speech/synthesize": {
      "post": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Speech Synthesize",
        "description": "Synthesize one already-normalised clause locally. Never leaves the machine.",
        "operationId": "speech_synthesize_v6_speech_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechSynthesizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "16-bit PCM WAV",
            "content": {
              "application/json": {
                "schema": {}
              },
              "audio/wav": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/competition/evidence": {
      "get": {
        "tags": [
          "v6 age-inclusive agent and competition evidence"
        ],
        "summary": "Competition Evidence",
        "operationId": "competition_evidence_v6_competition_evidence_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionEvidenceBoard"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InteractionProfile": {
        "properties": {
          "elder_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Elder Id"
          },
          "speech_rate": {
            "type": "number",
            "maximum": 1.2,
            "minimum": 0.6,
            "title": "Speech Rate",
            "default": 0.88
          },
          "verbosity": {
            "$ref": "#/components/schemas/VerbosityMode",
            "default": "gentle"
          },
          "max_options": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Max Options",
            "default": 3
          },
          "max_sentence_chars": {
            "type": "integer",
            "maximum": 90.0,
            "minimum": 16.0,
            "title": "Max Sentence Chars",
            "default": 42
          },
          "repeat_sensitive": {
            "type": "boolean",
            "title": "Repeat Sensitive",
            "default": true
          },
          "teach_back_high_risk": {
            "type": "boolean",
            "title": "Teach Back High Risk",
            "default": true
          },
          "font_scale": {
            "type": "number",
            "maximum": 1.8,
            "minimum": 1.0,
            "title": "Font Scale",
            "default": 1.25
          },
          "hearing_support": {
            "type": "boolean",
            "title": "Hearing Support",
            "default": false
          },
          "dialect_hint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialect Hint"
          },
          "family_id": {
            "type": "string",
            "title": "Family Id"
          },
          "updated_by": {
            "type": "string",
            "title": "Updated By"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "elder_id",
          "family_id",
          "updated_by",
          "updated_at",
          "version"
        ],
        "title": "InteractionProfile"
      },
...
      "SpeechSynthesizeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Text"
          },
          "speed": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.5,
            "title": "Speed",
            "default": 1.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SpeechSynthesizeRequest"
      },
...
      "TaskGlassBoxRequest": {
        "properties": {
          "heard_text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Heard Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "heard_text"
        ],
        "title": "TaskGlassBoxRequest"
      },
...
      "VerbosityMode": {
        "type": "string",
        "enum": [
          "concise",
          "standard",
          "gentle"
        ],
        "title": "VerbosityMode"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}